requirements.txt
line-bot-sdk==3.5.0
flask==2.2.5
requests==2.31.0